requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.5
aiosqlite==0.19.0
google-auth==2.29.0
requests==2.31.0
greenlet==3.0.3 